+++ examples/sfml/main.rs
@@ -13,11 +13,11 @@ use sfml::{
     window::{ContextSettings, Key, Style},
 };
 use std::{env, path::PathBuf, time::Duration};
-use tiled::{FilesystemResourceCache, FiniteTileLayer, Map};
+use tiled::{FiniteTileLayer, Loader, Map};
 use tilesheet::Tilesheet;
 
 /// A path to the map to display.
-const MAP_PATH: &'static str = "assets/tiled_base64_external.tmx";
+const MAP_PATH: &str = "assets/tiled_base64_external.tmx";
 
 /// A [Map] wrapper which also contains graphical information such as the tileset texture or the layer meshes.
 ///
@@ -85,8 +85,8 @@ impl Drawable for Level {
         target: &mut dyn RenderTarget,
         states: &sfml::graphics::RenderStates<'texture, 'shader, 'shader_texture>,
     ) {
-        let mut states = states.clone();
-        states.set_texture(Some(&self.tilesheet.texture()));
+        let mut states = *states;
+        states.set_texture(Some(self.tilesheet.texture()));
         for mesh in self.layers.iter() {
             target.draw_with_renderstates(mesh, &states);
         }
@@ -94,17 +94,17 @@ impl Drawable for Level {
 }
 
 fn main() {
-    let mut cache = FilesystemResourceCache::new();
+    let mut loader = Loader::new();
 
-    let map = Map::parse_file(
-        PathBuf::from(
-            env::var("CARGO_MANIFEST_DIR")
-                .expect("To run the example, use `cargo run --example sfml`"),
+    let map = loader
+        .load_tmx_map(
+            PathBuf::from(
+                env::var("CARGO_MANIFEST_DIR")
+                    .expect("To run the example, use `cargo run --example sfml`"),
+            )
+            .join(MAP_PATH),
         )
-        .join(MAP_PATH),
-        &mut cache,
-    )
-    .unwrap();
+        .unwrap();
     let level = Level::from_map(map);
 
     let mut window = create_window();
@@ -114,9 +114,8 @@ fn main() {
     loop {
         while let Some(event) = window.poll_event() {
             use sfml::window::Event;
-            match event {
-                Event::Closed => return,
-                _ => (),
+            if event == Event::Closed {
+                return;
             }
         }