+++ src/layers/mod.rs
@@ -85,7 +85,7 @@ impl LayerData {
                 ("offsety", offset_y, |v:String| v.parse().ok()),
                 ("parallaxx", parallax_x, |v:String| v.parse().ok()),
                 ("parallaxy", parallax_y, |v:String| v.parse().ok()),
-                ("name", name, |v| Some(v)),
+                ("name", name, Some),
                 ("id", id, |v:String| v.parse().ok()),
             ]
         );