+++ src/layers/tile/infinite.rs
@@ -31,8 +31,8 @@ impl InfiniteTileLayerData {
         let (e, c) = get_attrs!(
             attrs,
             optionals: [
-                ("encoding", encoding, |v| Some(v)),
-                ("compression", compression, |v| Some(v)),
+                ("encoding", encoding, Some),
+                ("compression", compression, Some),
             ]
         );
 
@@ -58,7 +58,12 @@ impl InfiniteTileLayerData {
         Ok(Self { chunks })
     }
 
-    pub(crate) fn get_tile(&self, x: i32, y: i32) -> Option<&LayerTileData> {
+    /// Obtains the tile data present at the position given.
+    ///
+    /// If the position given is invalid or the position is empty, this function will return [`None`].
+    ///
+    /// If you want to get a [`Tile`](`crate::Tile`) instead, use [`InfiniteTileLayer::get_tile()`].
+    pub fn get_tile_data(&self, x: i32, y: i32) -> Option<&LayerTileData> {
         let chunk_pos = tile_to_chunk_pos(x, y);
         self.chunks
             .get(&chunk_pos)
@@ -162,7 +167,7 @@ impl<'map> InfiniteTileLayer<'map> {
     /// If the position is empty, this function will return [`None`].
     pub fn get_tile(&self, x: i32, y: i32) -> Option<LayerTile> {
         self.data
-            .get_tile(x, y)
-            .and_then(|data| Some(LayerTile::new(self.map, data)))
+            .get_tile_data(x, y)
+            .map(|data| LayerTile::new(self.map, data))
     }
 }